gcp/requirements.txt
tensorflow==2.10.1
google-cloud-storage==1.42.1
Pillow==9.2.0